+++ plexify/tests/mocks.py
@@ -20,6 +20,6 @@ class MockCliArgs:
     subtitles: List[str]
     plex_location: Union[str, Path]
     label: str
-    use_download_location_as_label: bool
+    use_download_folder_as_label: bool
     verbose: bool
     log: Union[str, Path]