+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2.2.0] - 2022-01-12
 ### Added
-- Added --use_download_location_as_label CLI switch. [#7](https://github.com/jprevc/plexify/issues/7)
+- Added --use_download_folder_as_label CLI switch. [#7](https://github.com/jprevc/plexify/issues/7)
 
 ### Fixed
 - Wrongly determined output location, when media file contains '[ ]' [#8](https://github.com/jprevc/plexify/issues/8)